+++ src/main/java/util/Dao.java
@@ -4,8 +4,8 @@ public interface Dao<T> {
 
     public void create(T object);
 
-    public void update(int id, T object);
+    public void update(long id, T object);
 
-    public void delete(int id);
+    public void delete(long id);
 
 }
